대박 가즈아!!!!!!1231231

+++ lib/model/LottoStat.dart
@@ -1,5 +1,5 @@
 class LottoStat{
-  double lottoId = 0;
+  int lottoId = 0;
   int turn = 0;
   String oddAndEvenRate = "";
   int numberSum = 0;
